+++ src/main/java/com/limhaekyu/eco_rich_backend/dto/UpdateEmployeeInfoDto.java
@@ -16,5 +16,5 @@ public class UpdateEmployeeInfoDto {
     private String phoneNumber;
     private BigDecimal salary;
     private BigDecimal commissionPct;
-    private int departmentId;
+    private Long departmentId;
 }
